Fix duplicate correlation surfaces (#452)
Resolves https://github.com/tqec/tqec/issues/441.

I added a test which is sensitive to this bug. Thereby I also refactored
the (test) code a little bit.

I solved the bug by initializing the frontier with a fresh instantiation
of the leaf without the label. This solution fits best with the rest of
the code because this is actually done too when adding new nodes to the
frontier. It also makes sense in regard of the fact that the span just
reuses `ZXNode` in a way where a label has actually no meaning.