Support taking PyZX `GraphS` with 3D positions as inputs (#465)
This PR adds interoperability code for `PyZX`. The ultimate goal is to
completely remove the existing `ZXGraph` data structure, which will be
addressed in follow-up PRs.

In this PR, the `PositionedZX` data structure is introduced as a
container class for `GraphS` from `PyZX`, along with attached 3D
positions for each vertex. Additionally, the code from
`tqec/computation/conversion.py` has been adapted into
`tqec/interop/pyzx/synthesis/positioned.py` to support the conversion of
a `PositionedZX` into a `BlockGraph`. In the future, additional block
synthesis strategies should be added under the same directory.

---------

Co-authored-by: Adrien Suau <adrien.suau+github@protonmail.com>
Co-authored-by: Adrien Suau <nelimee.dev+github@protonmail.com>