Remove `ZXGraph` data structure from the code base (#466)
This is a follow-up PR of #465 and tries to remove `ZXGraph` data
structure completely.

Currently, the algorithm to find correlation surface is kept and adapted
to take a `GraphS` as input, instead of utilizing `compute_pauli_webs`
from `PyZX`. That’s because:

1. I don't fully understand the algorithm in `compute_pauli_webs`.
2. `compute_pauli_webs` requires specifying `inputs`/`outputs` boundary
vertices in ZX graph and does not support computing on no-port graphs.

---------

Co-authored-by: Adrien Suau <nelimee.dev+github@protonmail.com>
Co-authored-by: Adrien Suau <adrien.suau+github@protonmail.com>