Refactor `BlockGraph` and improve ergonomics (#467)
This PR removes the `ComputationGraph` abstract class, as `ZXGraph` has
been removed, eliminating the need for a base graph. Additionally,
`BlockGraph` has been refactored a bit to improve ergonomics:

- `add_cube` now accepts `position`, `kind`, and `label` directly
instead of requiring a pre-constructed `Cube`. The `kind` parameter can
also be a `str`, eliminating the need for frequent `ZXCube.from(...)`
conversions. And it now returns the position of the cube that can be
further used as handle in the `add_pipe` method.
- `add_pipe` now takes two cube positions and raises an exception if no
cube exists at either position. The `kind` parameter also supports `str`
for improved usability.

Additionally, this PR implements `compose` method on `BlockGraph`, which
should close #404.

Other minor changes can be found in the commit messages.

There are several documentation updates required based on this and the
last two PRs. I will address them in a separate PR in the coming days.

---------

Co-authored-by: Adrien Suau <nelimee.dev+github@protonmail.com>
Co-authored-by: Adrien Suau <adrien.suau+github@protonmail.com>