Introduce `tqec.compile.blocks` (#476)
This PR introduces the module `tqec.compile.blocks`.

The goal of this module is to replace `CompiledGraph` by a
data-structure (`TopologicalComputationGraph`) that will allow more
"scheduling schemes" to be returned by the `compile_block_graph`
function.

### Why is `CompiledGraph` not enough?

One way of implementing spatial junctions requires to alternate two sets
of plaquettes. This is currently impossible to encode using
`CompiledGraph`.

As a side note, `CompiledGraph` is inherently limited to computation
that can be represented using `Template`/`Plaquettes` instances. The
next goal(s) for `tqec` is to support such computations (e.g., Y-basis
state preparation / measurement).

### What is the replacement?

The architecture introduced in this PR revolves around the core concept
of a layer:

- a `BaseLayer` (and all its subclasses) represents the equivalent of
one round of the QEC code.
- a `BaseComposedLayer` (and all its subclasses) represents 2 or more
`BaseLayer` instances applied one after the other.

`BaseLayer` instances can be:

- `PlaquetteLayer` that uses the usual `Template` / `Plaquettes` combo
to represent a layer,
- `RawCircuitLayer` that uses a user-provided circuit construction
function to represent a layer,
- `LayoutLayer` that represents several `BaseLayer` instances executed
in parallel on a 2-dimensional grid.

`BaseComposedLayer` instances can be:

- `RepeatedLayer` that repeats a given number of times (that can scale
with `k`) an internal circuit that is represented either as a
`BaseLayer` instance or as a `BaseComposedLayer` instance,
- `SequencedLayers` that represents a sequence of layers that are
applied one after the other in time.

I think that this architecture is capable of representing most of the
quantum circuits we will care about in the near future.

Should fix #464.

----

*Note: this PR replaces #460 and is based on `main` instead of
`feat/spatial_junctions`.*

----

TODO list:

- [x] Make sure that layers from temporal pipes are z-aligned with
temporal boundary layers of cubes.
- [ ] Introduce detector computation to the layer approach.
- [ ] Introduce observable construction to the layer approach.

---------

Co-authored-by: Yiming Zhang <61700160+inmzhang@users.noreply.github.com>
Co-authored-by: Yilun Zhao <40353317+Zhaoyilunnn@users.noreply.github.com>