Update the description of label attribute (#514)
@inmzhang 

Clarified that any str can be used as a label, but must be unique:

The label of the cube. It’s mainly used for annotating the input/output
ports of the block graph. If the cube is a port, the label must be
non-empty **and unique within the block graph. The label can be any
string, but duplicate labels are not allowed**. Default is an empty
string.

---------

Co-authored-by: Yiming Zhang <61700160+inmzhang@users.noreply.github.com>