Representing computation as a tree, detectors and observables are annotations (#503)
This is a tentative PR to try to decouple as much as possible observable
and detector computation from the circuit-generation process.

The principle of this tentative is the following:

- merging the layers of a `TopologicalComputationGraph` now gives a
`LayerTree`,
- the `LayerTree` resulting from
`TopologicalComputationGraph.to_layer_tree()` has a `SequencedLayers`
instance as root, and each children of the root is also a
`SequencedLayers`, to keep something similar to the
`list[list[LayoutLayer | BaseComposedLayer]]`,
- the `LayerNode` instances composing the `LayerTree` are storing the
layer they represent, its children, and some annotations that, for the
moment, are related to observables and detectors. The exact typing of
each annotation is very much open. The intended goal is that annotations
should be appended at the end of the quantum circuit generated for a any
given node.

@inmzhang and @Zhaoyilunnn that are working on #501 and #502, here is
what I have in mind for your respective PRs, let me know if that makes
sense to you, or if you think that is not a good approach.

For #501, we "just" have to walk through the grand-children in the tree
and add observable annotation to them according to the provided
`AbstractObservable` instance.

For #502, we explore the computation tree using a DFS and we keep in
memory the last `N` `LayoutLayer` instances seen, `N` being the amount
of rounds in the past we want to look back at in order to compute
detectors. A special treatment will have to be performed on nodes
representing a `RepeatedLayer` (e.g., checking that the detectors inside
the loop are valid for the `N - 1` first rounds of the loop), but I
think this should be manageable.

If that idea seems reasonable to you, then we can likely make detector
and observables annotation independent of the actual circuit generation
(they will share the same scaling parameter `k`, but we will not have to
perform detector/observable computation *while* generating the circuit).

---------

Co-authored-by: Yiming Zhang <61700160+inmzhang@users.noreply.github.com>
Co-authored-by: Yilun Zhao <40353317+Zhaoyilunnn@users.noreply.github.com>