Correct space time volume definition (#543)
The definition of "spacetime volume" used throughout Austin's lectures
is "the number of non-port cubes the computation has".

Previously, it was defined as "the size of the bounding box of the block
graph":


https://github.com/tqec/tqec/blob/06420e1b53f227c2f1b6bcb18524f1e41a3c7d03/src/tqec/computation/block_graph.py#L113-L119

This PR renames it as `bounding_box_size` and add the method
`spacetime_volume` with the correct definition.