Fix layer approach (#549)
This PR fixes #539.

- [x] Allow adding a temporal pipe on a block with existing spatial
pipes,
- [x] Avoid requiring the same spatial shapes in `SequencedLayers`.

Note that the second issue was solved by simply delaying the "same
`scalable_shape`" check to when `SequencedLayers.scalable_shape` is
called, which is fine for the moment because we never call that property
after merging the layers, and the only supported way to get a
`LayoutLayer` is by merging the layers.

---------

Co-authored-by: Yilun Zhao <40353317+Zhaoyilunnn@users.noreply.github.com>