Fix cnot_all_observables.py (#568)
This fixes issue #566.

The code now runs. The output looks plausible. It is difficult to
confirm that it is the same as that output by the old compile pipeline
as there have been several commits since the deprecated compile script
was deleted.

My changes to the script are minimal. I have replaced the default
arguments of CSS_BLOCK_BUILDER and CSS_SUBSTITUTION_BUILDER in
compile_block_graph(), with the new default parameter of
FIXED_BULK_CONVENTION in the new compile_block_graph function. It would
be good if someone more familiar with the new pipeline could confirm
that the new default parameters are intended to produce the same result
as the old default parameters.