Introduce BlockGraph.relabel_cubes (#531)
See issue #527 (there will be a second PR for `fill_ports` but need to
have `relabel_cubes` solved first)

Relabel cubes searching the cubes by `position` or by `label` (allows to
relabel a group of cubes). It raises an exception if the name given is
been used in an existing port. I've tested it with:

```
from tqec import BlockGraph, Position3D

def main():    
    block_graph = BlockGraph.from_dae_file("slide67_open.dae")
    
    # let's see the list of cubes in the blockgraph
    print(f"Cubes:")
    for cube in block_graph.cubes:
        print(cube, cube.label)
    
    block_graph.relabel_cubes(Position3D(-1,0,0), "PortNew1")

    print(f"New Cubes:")
    for cube in block_graph.cubes:
        print(cube, cube.label)
    block_graph.relabel_cubes("Port1", "PortNew2")

    print(f"Cubes:")
    for cube in block_graph.cubes:
        print(cube, cube.label)

    block_graph.view_as_html("open_ports.html")

if __name__ == "__main__":
    main()
```