First version of SketchUp extension to import JSON files (#583)
Instructions to use the extension:

1. With the template.skp opened in SketchUp, save each element as a
component. They are going to be saved as zxx.skp, zxz.skp, etc. in
Components folder (save the pathname of that folder to use it in the
next step).
2. In line `COMPONENT_FOLDER ||=` replace the text with the pathname to
the `Components` folder in SketchUp.
3. Save the `tqec_blockgraph_loader.rb` in the `Plugins` folder of
SketchUp.
4. Now you can import a JSON file with the blockgraph and the
component's name won't be lost. It's possible to export the estructure
created in SketchUp as a DAE file and re-import it in TQEC.