Support json format when saving detector database to disk (#584)
Introduces a JSON exporter for the DetectorDatabase

This PR implement a new exporter to be able to avoid pickle when the 
serialized database does not come from a trusted source. 

The security concern comes from the fact that unpickling might lead 
to arbitrary code execution, which is highly not desirable when 
the unpickled file does not come from a trusted source.