Add concurrency option to docs workflow (#616)
This PR adds a concurrency option to the docs workflow such that if
there is a CI run in progress while a commit has been made, a user does
not have to manually cancel the workflow. This is helpful for the
`gh-pages` build as it often waits for the previous commit's build to be
deployed before building the docs for the latest commit.

The user gets a notification that looks like below. This was a result of
pushing an empty commit in
https://github.com/tqec/tqec/pull/616/commits/433fef9fe78d023c13b5f1c8de1ef43d71f5d915
right after the previous commit.


![image](https://github.com/user-attachments/assets/35909071-7af7-462e-a55f-e254ba29bd23)


https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/control-the-concurrency-of-workflows-and-jobs#example-using-concurrency-and-the-default-behavior