Improving parallelism when computing detectors (#595)
#482 points out that the process of generating stim circuit from graph
can be accelerated using parallel processing. This PR implements the
solution according to [this comment
](https://github.com/tqec/tqec/issues/482#issuecomment-2705827343)
and observations made in #595 when benchmarking.

For the moment, parallel processing is disabled if an existing non-empty
detector database is provided as synchronisation of the database across 
processes seems more costly than single-threaded performance. That
may be an idea for future improvements.

---------

Signed-off-by: zhao-yilun-zyl <dilved1014@gmail.com>
Co-authored-by: Adrien Suau <12374487+nelimee@users.noreply.github.com>