Use RPNG to build extended stabilizer (#651)
This PR is a first step toward fixing #631. 

It changes slightly how extended plaquettes are built, allowing more
freedom in the types of stabilizers that can be measured.

At the moment, debug information and in particular drawers are **NOT**
included. That means that extended plaquettes will not be drawn. That is
a regression I plan to fix later.

---------

Co-authored-by: Yiming Zhang <61700160+inmzhang@users.noreply.github.com>