fix description (#689)
A quick fix to the description of one of the functions for the spatial
horizontal hadamard. The function was behaving correctly, but the
description confusingly didn't match.

I thought it was clearer to push this separately, rather than wrap it
into my more complicated push implementing the other spatial hadamard
situations.