Set background color to white in HTML viewer (#733)
Changed the background color to white in the HTML Viewer for readability
(based on #673). The original issue sought to change only the options
text to have a white background, but when testing it with a couple
different themes, it didn't look right. I originally tried to get the
background color using the following code:
```
getComputedStyle(document.documentElement)
    .getPropertyValue('--vscode-editor-background')
    .trim()
    .toLowerCase();
```
I got the color, but choosing when to change the color of text (since VS
Code has such a wide assortment of color themes is quite difficult, so I
stuck to keeping the default background color to white.